refactor: move saveToJSONFile to fileUtils
